Converts Packet object to a regular Scala object
Tester using regular scala improves readability significantly and
makes sense as testing module can be ran in the scala language.

If it was a component it would have to follow all of the design
rules and that doesnt make much sense as it by nature does file
IO.

Furthermore instead of having a packet for tx and a packet for rx,
Packet will rx data and then check the recieved data to ensure
it matches the data it sent via rx. Reduces complexity and saves
memory space. Easily extendible to handle multiple data via an
ordered list